requirements.txt
requests==2.11.1
django-suit==0.2.22
Django==1.10.3
djangorestframework==3.5.3
django-cleanup
django-cors-headers